Cleaned up completion management
I removed refences for the parent client form the completion manager and moved the stream management  logic to the main file. I want the main file to handle the logic of passing text to TTS or the clipboard.
The completion manager is now more stateless which should make it easier when setting up LLM chains with multiple concurrent calls